+++ packages/clerk_flutter/lib/src/utils/default_caching_persistor/default_caching_persistor.dart
@@ -1,2 +1,2 @@
-export 'default_caching_persistor_io.dart'
+export 'default_caching_persistor_native.dart'
     if (dart.library.js_interop) 'default_caching_persistor_web.dart';
